server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
psutil==5.9.6
python-multipart==0.0.6

# AI/ML Dependencies for YOLOv8
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0

# Discord OAuth Authentication Dependencies
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
python-dateutil==2.8.2